apps/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
annotated-types==0.5.0
    # via pydantic
anyio==4.0.0
    # via
    #   httpcore
    #   starlette
build==1.0.0
    # via pip-tools
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   pip-tools
    #   uvicorn
exceptiongroup==1.1.3
    # via anyio
fastapi==0.103.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
packaging==23.1
    # via build
pip-tools==7.3.0
    # via -r requirements.in
pydantic==2.3.0
    # via fastapi
pydantic-core==2.6.3
    # via pydantic
pyproject-hooks==1.0.0
    # via build
python-decouple==3.8
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.27.0
    # via fastapi
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.7.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   uvicorn
uvicorn==0.23.2
    # via -r requirements.in
wheel==0.41.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
